attached_assets/requirements_1763002528231.txt
python-telegram-bot==22.5
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.2
openai==1.58.1
httpx==0.28.1
hijri-converter==2.3.1
aiosqlite==0.20.0
apscheduler==3.10.4
hijri-converter==2.3.1
httpx==0.28.1
openai==1.58.1
python-dotenv==1.0.1
python-telegram-bot==22.5
pytz==2024.2
telegram
